feat(condo): DOMA-7709 added resident payment alert
